Share memory in PutRequest in server
