+++ ambry-protocol/src/main/java/com.github.ambry.protocol/PutRequest.java
@@ -208,10 +208,12 @@ public class PutRequest extends RequestOrResponse {
       stream = new DataInputStream(crcInputStream);
       BlobId id = new BlobId(stream, map);
       BlobProperties properties = BlobPropertiesSerDe.getBlobPropertiesFromStream(stream);
-      ByteBuffer metadata = Utils.readIntBuffer(stream);
+      int mdSize = stream.readInt();
+      ByteBuffer metadata = Utils.readByteBufferFromCRCInputStream(crcInputStream, mdSize);
       BlobType blobType = BlobType.values()[stream.readShort()];
       long blobSize = stream.readLong();
-      ByteBufferInputStream blobStream = new ByteBufferInputStream(stream, (int) blobSize);
+      ByteBufferInputStream blobStream =
+          Utils.getByteBufferInputStreamFromCRCInputStream(crcInputStream, (int) blobSize);
       long computedCrc = crcInputStream.getValue();
       long receivedCrc = stream.readLong();
       if (computedCrc != receivedCrc) {
@@ -233,11 +235,13 @@ public class PutRequest extends RequestOrResponse {
       stream = new DataInputStream(crcInputStream);
       BlobId id = new BlobId(stream, map);
       BlobProperties properties = BlobPropertiesSerDe.getBlobPropertiesFromStream(stream);
-      ByteBuffer metadata = Utils.readIntBuffer(stream);
+      int mdSize = stream.readInt();
+      ByteBuffer metadata = Utils.readByteBufferFromCRCInputStream(crcInputStream, mdSize);
       BlobType blobType = BlobType.values()[stream.readShort()];
       ByteBuffer blobEncryptionKey = Utils.readShortBuffer(stream);
       long blobSize = stream.readLong();
-      ByteBufferInputStream blobStream = new ByteBufferInputStream(stream, (int) blobSize);
+      ByteBufferInputStream blobStream =
+          Utils.getByteBufferInputStreamFromCRCInputStream(crcInputStream, (int) blobSize);
       long computedCrc = crcInputStream.getValue();
       long receivedCrc = stream.readLong();
       if (computedCrc != receivedCrc) {
